Make DaoAuthorizable  and Plugin use a constructor.

+++ packages/contracts/contracts/plugin/examples/count/MultiplyHelper.sol
@@ -1,17 +1,19 @@
 // SPDX-License-Identifier: MIT
- 
+
 pragma solidity 0.8.10;
 
-// import "../../../core/plugin/PluginClones.sol";
+// import "../../../core/plugin/PluginCloneable.sol";
 import "../../../core/plugin/PluginUUPSUpgradeable.sol";
 
 // NON-Upgradable
-contract MultiplyHelper is PluginUUPSUpgradeable  {
-    
+contract MultiplyHelper is PluginUUPSUpgradeable {
     bytes32 public constant MULTIPLY_PERMISSION_ID = keccak256("MULTIPLY_PERMISSION");
 
-    function multiply(uint a, uint b) external auth(MULTIPLY_PERMISSION_ID) returns (uint) {
+    function multiply(uint256 a, uint256 b)
+        external
+        auth(MULTIPLY_PERMISSION_ID)
+        returns (uint256)
+    {
         return a * b;
     }
-
 }
